+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - HttpStorage deprecated: jsons were saved as sent by the server, now we save json with logic (saving ids and resources separately).
 - Service with `toServer()` and `fromServer()` functions. They execute before and after http request. Ideal for type conversions.
 - `JsonapiCore.duplicateResource(resouce, ...relationtypes)` return a duplication of resource. You can duplicate resources and, optionally, their relationships. (v0.6.16)
+- resource save() method return a promise.
 
 ## No more declaration file .d.ts